create PageTreeItem
